Migração da lógica de `Shelf.vue` para `Home.vue`
